+++ Project Files/Source/Console/console.cs
@@ -14938,6 +14938,13 @@ namespace PowerSDR
 
         public void SetAlexLPF(double freq)
         {
+            if (!mox && lpf_bypass)
+            {
+                JanusAudio.SetAlexLPFBits(0x10); // 6m LPF
+                SetupForm.rad6LPFled.Checked = true;
+                return;
+            }
+
             if (chkPower.Checked && alexpresent && SetupForm.radAlexManualCntl.Checked)
             {
                 if ((decimal)freq >= SetupForm.udAlex20mLPFStart.Value && // 30/20m LPF
@@ -22205,6 +22212,7 @@ namespace PowerSDR
                 vfo_lock = value;
                 bool enabled = !value;
                 txtVFOAFreq.Enabled = enabled;
+                txtVFOBFreq.Enabled = enabled;
                 radBand160.Enabled = enabled;
                 radBand80.Enabled = enabled;
                 radBand60.Enabled = enabled;
@@ -25104,6 +25112,7 @@ namespace PowerSDR
             }
             set
             {
+                if (vfo_lock || SetupForm == null) return;
                 value = Math.Max(0, value);
                 txtVFOBFreq.Text = value.ToString("f6");
                 txtVFOBFreq_LostFocus(this, EventArgs.Empty);
@@ -25557,6 +25566,24 @@ namespace PowerSDR
                 {
                     double freq = Double.Parse(txtVFOAFreq.Text);
                     SetAlexHPF(freq);
+                }
+            }
+        }
+
+        private bool lpf_bypass = false;
+        public bool LPFBypass
+        {
+            get { return lpf_bypass; }
+            set
+            {
+                lpf_bypass = value;
+                if (chkPower.Checked)
+                {
+                    double freq = Double.Parse(txtVFOAFreq.Text);
+                    if (mox) freq = tx_dds_freq_mhz;
+                    SetAlexLPF(freq);
+                    if (!initializing)
+                        txtVFOAFreq_LostFocus(this, EventArgs.Empty);
                 }
             }
         }