+++ ReleaseNotes.md
@@ -1,4 +1,4 @@
-# PowerSDR_mRX_PS 3.4.6.0 November 20, 2017
+# PowerSDR_mRX_PS 3.4.7.0 December 22, 2017
 
 
 # 3.2.22 (2015-1-24) 
@@ -341,3 +341,8 @@ Fixed a bug resulting in incorrect vertical display scale in transmit under cert
 
 # 3.4.6 (2017-11-20)
 - Corrected problem with blank waterfall display in Panafall mode.
+
+# 3.4.7 (2017-12-22)
+- Control added to force the LPF to the 6m/ByPass posistion during receive. Filters must be under manual control to use. (Setup=>General=>Ant/Filters=>LPF, HPF/LPF, BPF1)
+- VFO Lock correctly locks VFOA and VFOB.
+- TX Amplitude scaling adder to waterfall display. (Setup=>Display=>TX)